Refactored TestConfig creation for py.test based unit tests.
The method used before had multiple problems: config creation did
not use inheritance (as it does when used the wiki in production).
It also didn't do the post processing done in __init__ of Config
class. Also, it first created the request with some default config,
then created the TestConfig (but some stuff in request was not
initialized correctly)...

The new method now uses config inheritance in the same way as a
production wiki does. The TestConfig instance is created in RequestBase
__init__, it does not use sys.path any more to import the test
wikiconfig, but does an absolute import from MoinMoin._tests.wikiconfig.

TODO:
 * fix the failing test_pysupport.TestImportExisting.testExisting
 * auth tests are currently skipped because they need more work
 * find a better way than staticmethod(TestConfig)