mailimport: fix endless looping while trying to import a forwarded mail
This is just a fix to solve the endless looping / crash.

It needs further improvement, currently the From/To/... header data
of the forwarded mail is not processed at all and thus not visible
later in the wiki. The body of it is visible, though, but just appears
appended to the main mail body.