MoinMoin.util.filesys: disable usage of dircache, deprecate dc* functions (details below)
The dircache stdlib module can't work for fast updates of directories.

It uses an mtime check on the directory to determine whether it has changed
and returns cached stuff if it thinks it didn't change.
This will just go wrong, if you update the directory faster than directory mtime
granularity of the filesystem (the mtime might not change then).

Thus, for ext3, it breaks if you update faster than 1s (that broke some of our
unit tests).

For FAT32 we had found earlier that it is completely broken (IIRC because FAT32
does not support directory mtime). We had it disabled therefore completely on
win32 platform.

For now, the dc* functions are emitting DeprecationWarnings and either doing
nothing or calling the uncached os.listdir() function.

Next, we'll fix the callers and later, dc* will get removed completely.