moin account resetpw - added options --all-users and --notify
--all-users is to reset the password of all (even disabled) users (use some
random and secret password for this). Note: if -q (quiet) flag is not given,
it will show some progress indication with the current count / total count
and uid.

--notify can be used to notify the user of the account via e-mail and send a
password reset link. when used with --all-users, all users that have an e-mail
address and are not disabled will be notified by e-mail (caution: depending on
your wiki's user count, this can generate a huge count of e-mails and take
quite a while to be processed).

refactored the code to use exceptions, move some code to user module