introduce a new caching scope "userdir"
Needed this because none of the existing scopes is usable for caching data
related to the data that is contained in the user_dir (user profiles).

This is a problem for shared user_dir scenarios, when multiple wikis share
the same user_dir. This is common in farm setups with single-sign-on.

"farm" scope can't be used though as not all farm wikis are required to share
the users. Also, non-farm setups may also have shared user_dir.

Thus, best is to have this new caching scope (and even have it stored inside
the user_dir directory, so it is automatically the "right" directory without
needing further configuration). Having the "cache" directory inside the
user_dir is no issue as the user profiles follow a specific naming pattern and
we filter for that pattern.