replace slow get_by_filter() calls by faster _getUserIdByKey calls
get_by_filter is O(N) with N being the amount of users.
it reads all user profiles from disk, so it is very slow if you have many.

_getUserIdByKey uses a cached dictionary with direct lookup search value -> userid

implement optional case-insensitive lookup for _getUserIdByKey

use lower-cased search value and lower-cased dict keys to support the case-insensitive

the lowercase support dicts are only built / kept in memory and not cached to disk