added cfg.actions_superuser to restrict actions to superusers
include 'newaccount' action in the default list for it because spam
bots like to create tons of user accounts.

Modify superuser's "Settings" -> "Switch User" functionality to allow
creating a new user (or switching to a user by typing the username as
an alternative to selecting the name from the list).

The minimum a superuser should fill out for a new account is the email
address.

It is not required to set a password, the user can just use the "forgot
password" functionality. Before the password is set, the account can not
be used to log in.