workaround werkzeug errors='fallback:...' regression, fixes #37
werkzeug had a documented feature of supporting a fallback coding
for decoding strings until yesterday, but it was broken due to removal
of the supporting code since many years.

only moin users with iso-8859-1 contents in their event-log noticed
that, for the usual utf-8 everything was fine.

i added a simulation of that behaviour now to our code, so hopefully
everything works now.