catch indexer error for too long names, fixes #57
xapian raises an exception if wikiname:pagename/attachmentname is > 254
chars.

this used to abort the whole indexing run, but now we catch and log
this exception so that only 1 item is affected, but the indexing run
continues for the rest of items.

also improved indexer logging to use %r so logging never crashes due to
encoding issues for non-ascii page or attachment names.